@@ -24,11 +24,10 @@
         }
     ],
     "require": {
-        "php": ">=8.0.2",
+        "php": ">=8.1.0",
         "symfony/deprecation-contracts": "^2.5|^3",
         "symfony/polyfill-mbstring": "^1.3",
-        "symfony/polyfill-ctype": "^1.8",
-        "symfony/polyfill-php81": "^1.29"
+        "symfony/polyfill-ctype": "^1.8"
     },
     "require-dev": {
         "symfony/phpunit-bridge": "^5.4.9|^6.4|^7.0",
