@@ -87,19 +87,12 @@ class DefaultExceptionHtmlSubscriber extends HttpExceptionSubscriberBase {
    *   The event to process.
    */
   public function on4xx(ExceptionEvent $event) {
-    if (($exception = $event->getThrowable()) && $exception instanceof HttpExceptionInterface) {
+    // Avoid making a subrequest for 400 errors because the same conditions that
+    // caused the 400 error could also happen in the subrequest. This allows 400
+    // exceptions to fall through to FinalExceptionSubscriber::on4xx.
+    if (($exception = $event->getThrowable()) && $exception instanceof HttpExceptionInterface && $exception->getStatusCode() > 400) {
       $this->makeSubrequest($event, '/system/4xx', $exception->getStatusCode());
     }
-  }
-
-  /**
-   * Handles a 400 error for HTML.
-   *
-   * @param \Symfony\Component\HttpKernel\Event\ExceptionEvent $event
-   *   The event to process.
-   */
-  public function on400(ExceptionEvent $event): void {
-    throw $event->getThrowable();
   }
 
   /**
