@@ -24,7 +24,8 @@ class FileAudioFormatterTest extends FileMediaFormatterTestBase {
    * @dataProvider dataProvider
    */
   public function testRender($tag_count, $formatter_settings): void {
-    $field_config = $this->createMediaField('file_audio', 'mp3', $formatter_settings);
+    // Create a file field that accepts .mp3 and an unknown file extension.
+    $field_config = $this->createMediaField('file_audio', 'unknown-extension, mp3', $formatter_settings);
 
     file_put_contents('public://file.mp3', str_repeat('t', 10));
     $file1 = File::create([
