@@ -441,7 +441,7 @@ class VariationCacheTest extends UnitTestCase {
     // a cache redirect should always be present on the redirect itself. In this
     // example, the final cache redirect should be for 'A,B:foo,B'.
     $this->expectException(\LogicException::class);
-    $this->expectExceptionMessage('Trying to overwrite a cache redirect with one that has nothing in common, old one at address "house.type" was pointing to "garden.type:zen", new one points to "garden.type".');
+    $this->expectExceptionMessage('Trying to overwrite a cache redirect for "your:housing:situation:ht.house" with one that has nothing in common, old one at address "house.type" was pointing to "garden.type:zen", new one points to "garden.type".');
 
     $this->housingType = 'house';
     $house_cacheability = (new CacheableMetadata())
@@ -484,7 +484,7 @@ class VariationCacheTest extends UnitTestCase {
     // previous redirects should always be present on the next redirect or item
     // you're trying to store.
     $this->expectException(\LogicException::class);
-    $this->expectExceptionMessage('Trying to overwrite a cache redirect with one that has nothing in common, old one at address "house.type" was pointing to "garden.type", new one points to "house.orientation".');
+    $this->expectExceptionMessage('Trying to overwrite a cache redirect for "your:housing:situation:ht.house" with one that has nothing in common, old one at address "house.type" was pointing to "garden.type", new one points to "house.orientation".');
 
     $this->housingType = 'house';
     $house_cacheability = (new CacheableMetadata())
@@ -523,7 +523,7 @@ class VariationCacheTest extends UnitTestCase {
     }, E_USER_WARNING);
 
     $this->expectException(\LogicException::class);
-    $this->expectExceptionMessage('Trying to overwrite a cache redirect with one that has nothing in common, old one at address "house.type, garden.type" was pointing to "house.orientation", new one points to "solar.type".');
+    $this->expectExceptionMessage('Trying to overwrite a cache redirect for "your:housing:situation:gt.garden:ht.house" with one that has nothing in common, old one at address "house.type, garden.type" was pointing to "house.orientation", new one points to "solar.type".');
 
     $this->housingType = 'house';
     $house_cacheability = (new CacheableMetadata())
@@ -572,7 +572,7 @@ class VariationCacheTest extends UnitTestCase {
     }, E_USER_WARNING);
 
     $this->expectException(\LogicException::class);
-    $this->expectExceptionMessage('Trying to overwrite a cache redirect with one that has nothing in common, old one at address "house.type, garden.type" was pointing to "house.orientation", new one points to "solar.type".');
+    $this->expectExceptionMessage('Trying to overwrite a cache redirect for "your:housing:situation:gt.garden:ht.house" with one that has nothing in common, old one at address "house.type, garden.type" was pointing to "house.orientation", new one points to "solar.type".');
 
     $this->housingType = 'house';
     $house_cacheability = (new CacheableMetadata())
