@@ -202,7 +202,7 @@ class OpenIDConnectSettingsForm extends ConfigFormBase {
     $form['role_mappings'] = [
       '#title' => 'EXPERIMENTAL - ' . $this->t('User role mapping'),
       '#type' => 'fieldset',
-      '#description' => $this->t('For each Drupal role, provide the sets of equivalent external groups, separated by spaces. A user belonging to one of the provided groups will be assigned the configured Drupal role.') .
+      '#description' => $this->t('For each Drupal role, provide the sets of equivalent external groups, separated by spaces. If the external group includes spaces, be sure to wrap the group name in double quotes (e.g. "content editor"). A user belonging to one of the provided groups will be assigned the configured Drupal role.') .
                         $this->t("<br/><strong>Note:</strong> The module will not update user roles with no mapped external groups. If all mappings to one of the roles are removed, users will keep that role until it is removed in the Drupal user administration."),
       '#tree' => TRUE,
     ];
