@@ -45,7 +45,7 @@ class SchedulerRulesConditionsTest extends SchedulerBrowserTestBase {
   /**
    * Tests the conditions for whether an entity type is enabled for Scheduler.
    *
-   * @dataProvider dataStandardEntityTypes()
+   * @dataProvider dataStandardEntityTypes
    */
   public function testEntityTypeEnabledConditions($entityTypeId, $bundle) {
 
@@ -171,7 +171,7 @@ class SchedulerRulesConditionsTest extends SchedulerBrowserTestBase {
   /**
    * Tests the conditions for whether an entity is scheduled.
    *
-   * @dataProvider dataStandardEntityTypes()
+   * @dataProvider dataStandardEntityTypes
    */
   public function testEntityIsScheduledConditions($entityTypeId, $bundle) {
     // The legacy rules condition ids for nodes remain as:
