@@ -41,10 +41,14 @@ class UnPublishStateConstraintValidator extends ConstraintValidatorBase {
     // transition from the set published state to the un-publish state is
     // a valid transition.
     if ($publish_state && !$this->isValidTransition($entity, $publish_state, $unpublish_state)) {
+      $workflow_type = $this->getEntityWorkflowType($entity);
+      $mod_state_label = $workflow_type->hasState($moderation_state) ? $workflow_type->getState($moderation_state)->label() : 'None';
+      $unpub_state_label = $workflow_type->hasState($unpublish_state) ? $workflow_type->getState($unpublish_state)->label() : 'None';
+      $pub_state_label = $workflow_type->hasState($publish_state) ? $workflow_type->getState($publish_state)->label() : 'None';
       $this->context
         ->buildViolation($constraint->invalidPublishToUnPublishTransitionMessage, [
-          '%publish_state' => $publish_state,
-          '%unpublish_state' => $unpublish_state,
+          '%publish_state' => $pub_state_label,
+          '%unpublish_state' => $unpub_state_label,
         ])
         ->atPath('publish_state')
         ->addViolation();
@@ -54,10 +58,14 @@ class UnPublishStateConstraintValidator extends ConstraintValidatorBase {
     // the un-publish state is a valid transition based on the entity's
     // current moderation state.
     if (!$publish_state && !$this->isValidTransition($entity, $moderation_state, $unpublish_state)) {
+      $workflow_type = $this->getEntityWorkflowType($entity);
+      $mod_state_label = $workflow_type->hasState($moderation_state) ? $workflow_type->getState($moderation_state)->label() : 'None';
+      $unpub_state_label = $workflow_type->hasState($unpublish_state) ? $workflow_type->getState($unpublish_state)->label() : 'None';
+      $pub_state_label = $workflow_type->hasState($publish_state) ? $workflow_type->getState($publish_state)->label() : 'None';
       $this->context
         ->buildViolation($constraint->invalidUnPublishTransitionMessage, [
-          '%unpublish_state' => $unpublish_state,
-          '%content_state' => $moderation_state,
+          '%unpublish_state' => $unpub_state_label,
+          '%content_state' => $mod_state_label,
         ])
         ->atPath('publish_state')
         ->addViolation();
