@@ -12,7 +12,7 @@ class EventsTest extends SchedulerContentModerationBrowserTestBase {
   /**
    * Tests the PUBLISH_IMMEDIATELY Scheduler event subscriber.
    *
-   * @dataProvider dataEntityTypes()
+   * @dataProvider dataEntityTypes
    */
   public function testEvents($entityTypeId, $bundle): void {
     $this->drupalLogin($entityTypeId == 'media' ? $this->schedulerMediaUser : $this->schedulerUser);
