@@ -2,11 +2,11 @@
 
 namespace Drupal\Tests\scheduler_content_moderation_integration\Functional;
 
-use Drupal\commerce_product\Entity\ProductType;
 use Drupal\Tests\BrowserTestBase;
 use Drupal\Tests\content_moderation\Traits\ContentModerationTestTrait;
 use Drupal\Tests\scheduler\Traits\SchedulerMediaSetupTrait;
 use Drupal\Tests\scheduler\Traits\SchedulerSetupTrait;
+use Drupal\commerce_product\Entity\ProductType;
 
 /**
  * Base class from which all functional browser tests can be extended.
